+++ punchbowl/level2/merge.py
@@ -20,7 +20,6 @@ def merge_many_polarized_task(data: list[NDCube | None], trefoil_wcs: WCS) -> ND
             reprojected_uncertainties = np.stack([d.uncertainty.array for d in selected_images], axis=-1)
             reprojected_uncertainties[reprojected_uncertainties <= 0] = np.inf
             reprojected_uncertainties[np.isinf(reprojected_uncertainties)] = 1E64
-            reprojected_uncertainties[reprojected_data <= 0] = np.inf
 
             reprojected_weights = 1 / np.square(reprojected_uncertainties)
 
@@ -32,9 +31,11 @@ def merge_many_polarized_task(data: list[NDCube | None], trefoil_wcs: WCS) -> ND
             trefoil_uncertainty_layers.append(np.zeros((4096, 4096))-999)
 
     output_meta = NormalizedMetadata.load_template("PTM", "2")
-    output_dateobs = average_datetime([d.meta.datetime for d in data if d is not None]).isoformat()
-    output_datebeg = min([d.meta.datetime for d in data if d is not None]).isoformat()
-    output_dateend = max([d.meta.datetime for d in data if d is not None]).isoformat()
+    output_dateobs = average_datetime(
+        [d.meta.datetime for d in data if d is not None],
+    ).strftime("%Y-%m-%dT%H:%M:%S.%f")[:-3]
+    output_datebeg = min([d.meta.datetime for d in data if d is not None]).strftime("%Y-%m-%dT%H:%M:%S.%f")[:-3]
+    output_dateend = max([d.meta.datetime for d in data if d is not None]).strftime("%Y-%m-%dT%H:%M:%S.%f")[:-3]
     output_meta["DATE-OBS"] = output_dateobs
     output_meta["DATE-BEG"] = output_datebeg
     output_meta["DATE-END"] = output_dateend
@@ -47,7 +48,8 @@ def merge_many_polarized_task(data: list[NDCube | None], trefoil_wcs: WCS) -> ND
     )
 
 @punch_task
-def merge_many_clear_task(data: list[NDCube | None], trefoil_wcs: WCS, level: str = "2") -> NDCube:
+def merge_many_clear_task(
+        data: list[NDCube | None], trefoil_wcs: WCS, level: str = "2", maintain_nans: bool = False) -> NDCube:
     """Merge many task and carefully combine uncertainties."""
     trefoil_data_layers, trefoil_uncertainty_layers = [], []
     selected_images = [d for d in data if d is not None]
@@ -56,14 +58,16 @@ def merge_many_clear_task(data: list[NDCube | None], trefoil_wcs: WCS, level: st
         reprojected_data = np.stack([d.data for d in selected_images], axis=-1)
         reprojected_uncertainties = np.stack([d.uncertainty.array for d in selected_images], axis=-1)
         reprojected_uncertainties[reprojected_uncertainties <= 0] = np.inf
-        reprojected_uncertainties[np.isinf(reprojected_uncertainties)] = 1E64
-        reprojected_uncertainties[reprojected_data <= 0] = np.inf
+        reprojected_uncertainties[~np.isfinite(reprojected_uncertainties)] = 1E64
 
         reprojected_weights = 1/np.square(reprojected_uncertainties)
 
         trefoil_data_layers.append(np.nansum(reprojected_data * reprojected_weights, axis=-1) /
                                    np.nansum(reprojected_weights, axis=-1))
         trefoil_uncertainty_layers.append(1/np.sqrt(np.nansum(reprojected_weights, axis=-1)))
+        if maintain_nans:
+            was_nan = np.all(np.isnan(reprojected_data), axis=-1)
+            trefoil_data_layers[0][was_nan] = np.nan
     else:
         trefoil_data_layers.append(np.zeros((4096, 4096)))
         trefoil_uncertainty_layers.append(np.zeros((4096, 4096))-999)